requirements.txt
gunicorn==23.0.0
tensorflow==2.18.0
Flask==3.1.0
pillow==11.1.0
flask-sqlalchemy==3.1.1 
bcrypt==4.2.1
gevent==23.9.1  
uvicorn==0.34.0
uvicorn_worker==0.3.0
nextcord>=2.0.0,<3
python-dotenv==0.19.0
opencv-python==4.11.0.86